⬆️ Use Nedryland 8.0.0 and Nedryglot 1.0.0
Firm is now also a nix flake, which means the version of nixpkgs can be
controlled from the user side. Nedryland's check apps are forward and
combined with Firm's binaries, for easy use with `nix run .#bendini` for
example.

Use rust from nixpkgs for stable and oxalica for nightly.
Added a rust extension to provide:
- base.languages.rust.withWasi
- base.languages.rust.nightly
- base.languages.rust.nightly.withWasi
- base.languages.rust.nativeTools (with windows cross target)

Update CI to use flakes and use workflows and actions from Nedryland.

+++ libraries/rust/firm-rust/src/lib.rs
@@ -88,15 +88,15 @@ fn _map_attachment<S: AsRef<str> + std::fmt::Display>(
         &mut attachment_path_bytes_len as *mut usize
     ))?;
 
-    let mut attachment_path_buffer = Vec::with_capacity(attachment_path_bytes_len as usize);
+    let mut attachment_path_buffer = Vec::with_capacity(attachment_path_bytes_len);
     host_call!(raw::map_attachment(
         attachment_name.as_ref().as_ptr(),
         attachment_name.as_ref().as_bytes().len(),
         unpack as u8,
         attachment_path_buffer.as_mut_ptr(),
-        attachment_path_bytes_len as usize,
+        attachment_path_bytes_len,
     ))?;
-    unsafe { attachment_path_buffer.set_len(attachment_path_bytes_len as usize) };
+    unsafe { attachment_path_buffer.set_len(attachment_path_bytes_len) };
 
     Ok(PathBuf::from(String::from_utf8(attachment_path_buffer)?))
 }
@@ -135,15 +135,15 @@ fn _map_attachment_from_descriptor(
         &mut attachment_path_bytes_len as *mut usize
     ))?;
 
-    let mut attachment_path_buffer = Vec::with_capacity(attachment_path_bytes_len as usize);
+    let mut attachment_path_buffer = Vec::with_capacity(attachment_path_bytes_len);
     host_call!(raw::map_attachment_from_descriptor(
         value.as_ptr(),
         value.len(),
         unpack as u8,
         attachment_path_buffer.as_mut_ptr(),
-        attachment_path_bytes_len as usize,
+        attachment_path_bytes_len,
     ))?;
-    unsafe { attachment_path_buffer.set_len(attachment_path_bytes_len as usize) };
+    unsafe { attachment_path_buffer.set_len(attachment_path_bytes_len) };
 
     Ok(PathBuf::from(String::from_utf8(attachment_path_buffer)?))
 }
